pyproject.toml
[project]
name = "UniversalKGbuilder"
version = "0.1.0"
description = "Open Source Chemical Knowledge Graph Construction Pipeline"
readme = "README.md"
requires-python = "==3.11.*"
license = {text = "MIT"}
authors = [
    {name = "FlyWill", email = "WillingXun@outlook.com"}
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "pyyaml>=6.0",
    "chardet>=4.0.0",
    "openpyxl>=3.0.0",
    "psutil>=5.8.0",
    "requests>=2.26.0",
    "tqdm>=4.60.0",
    "openai>=1.0.0",
    "dashscope>=1.14.0",
    "google-generativeai>=0.3.0",
    "streamlit>=1.28.0",
    "watchdog>=3.0.0",
    "PyQt5>=5.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0"
]



